requirements.txt
-f /usr/share/pip-wheels
Django==2.2.4
mysqlclient==1.4.4
Pillow==6.2.1
PyMySQL==0.9.3
pytz==2019.2
sqlparse==0.3.0
